lets deploy this much for now

+++ src/components/Post.jsx
@@ -4,9 +4,16 @@ import { useDispatch } from 'react-redux';
 import { deletePost } from '../slices/postSlice';
 
 const Post = React.memo(({ post }) => {
-    const dispatch=useDispatch();
-    const deleteHandler=()=>{
-      dispatch(deletePost(post));
+    const dispatch = useDispatch();
+    const deleteHandler = () => {
+        const isConfirmed = confirm("Are you sure you want to delete this post?");
+        if (isConfirmed) {
+
+            dispatch(deletePost(post));
+            alert("Post deleted");
+        }else{
+            return;
+        }
     }
     return (
         <div
@@ -17,23 +24,34 @@ const Post = React.memo(({ post }) => {
                 <em>Date: 20/04/2081</em>
                 <p className='max-h-24 overflow-hidden'>{post.blogContent}</p>
             </Link>
-            <div className=' flex flex-row justify-end '>
+            <div className=' flex flex-row justify-between '>
+                <div>
 
-                <div className='mr-6'>
-                    (3)<i className="fa fa-comment postBadge hover:bg-transparent text-gray-600 cursor-default" aria-hidden="true"></i>
+                    <div className='mr-6'>
+                        (3)<i className="fa fa-comment postBadge hover:bg-transparent text-gray-600 cursor-default" aria-hidden="true"></i>
+                    </div>
                 </div>
-                <div className='' onClick={deleteHandler}>
-                    <i className="fa fa-trash  postBadge " aria-hidden="true"></i>
+                <div className='flex flex-row gap-x-4'>
+                    <div className='' onClick={deleteHandler}>
+                        <i className="fa fa-trash  postBadge " aria-hidden="true"></i>
+                    </div>
+                    <div className=''>
+                        <i className="fa fa-eye postBadge" aria-hidden="true"></i>
+                    </div>
+                    <Link to="/create" state={{ post }}>
+                        <i className="fas fa-edit postBadge"></i>
+                    </Link>
                 </div>
-                <div className=''>
-                    <i className="fa fa-eye postBadge" aria-hidden="true"></i>
-                </div>
-                <Link to="/create" state={{post}}>
-                <i className="fas fa-edit postBadge"></i>
-                </Link>
+
             </div>
         </div>
     )
 });
 
 export default Post;
+//Things left
+//1. either make post store in localstorage or backend
+//2. work on comments part (make them count number appear in posts too)
+//3. work on preview part
+//4. work on responsiveness
+//5. work on user profile
